+++ terminusdb_client/woqlquery/woql_query.py
@@ -384,6 +384,7 @@ class WOQLQuery:
             "range": "rdfs:range",
             "domain": "rdfs:domain",
             "subClassOf": "rdfs:subClassOf",
+            "boolean": "xsd:boolean",
             "string": "xsd:string",
             "integer": "xsd:integer",
             "decimal": "xsd:decimal",
@@ -756,6 +757,12 @@ class WOQLQuery:
 
     def string(self, input_str):
         return {"@type": "xsd:string", "@value": input_str}
+
+    def boolean(self, input_bool):
+        if input_bool:
+            return {"@type": "xsd:boolean", "@value": True}
+        else:
+            return {"@type": "xsd:boolean", "@value": False}
 
     def literal(self, input_val, input_type):
         if ":" not in input_type: